+++ CHANGELOG.md
@@ -4,6 +4,7 @@
   - Removed _Kills % in Stats display_ setting in favor of them
 - **Minimap zooming** (keyboard only)
 - Toggle to **disable the Killough-face easter egg**
+- Toggle to **make attackers face fuzzy targets straight**
 
 ## Changes